Add files via upload
This is a desktop calculator app, written in java, and JavaFx. I used the Eclipse editor and the SceneBuilder for designing, 
It contains a variety of mathematical operators like the sine, cosine, tangent, logarithm, natural logarithm, exponent, addition, multiplication, division, subtraction and others. 
 
Scene Builder allows you to easily layout JavaFX UI controls, charts, shapes, and containers, so
that you can quickly design user interfaces. It has the following advantages:
 It provides a clear separation of GUI from logic/controller.
 Rapid scene development / mock up
 It reduces the effort required to build user interfaces considerably. So time and
energy is saved.


This project has been divided into 4 files:

 The Main.java file: This file contains the main function, which has the launch() method.
This method is a static method located in the Application class. This method launches the
JavaFX runtime and your JavaFX application. This file also contains the Start class,
which when invoked displays the Graphical User Interface (GUI) to the screen.

 The Main.fxml file: FXML is an XML-based user interface markup language created by
Oracle Corporation for defining the user interface of a JavaFX application. This file
stores the user interface design using XML tags.

 The MainController.java file: This file is the controller of the main.fxml file. It serves as
the Logic behind the design. It listens to Action events and performs certain functions in
response to these events.
It contains a class called MainController, which has the following properties: Label
result, Label input, float number1, float number2, String operator, boolean start.
- The input and result labels are specific IDs used to access the Labels which
display the input from the user, and the results of the computations respectively.
- The number1 and number2 floats are variables used to store the value of the
operands.
The string operator is used to store the value of the operator.
- The Boolean start is a variable used as a flag to detect the start and end of a
computation. When it’s value is TRUE, it means a computation is ongoing, but
when it’s value is FALSE, it means no computations are being processed by the
calculator.
This file also contains some methods which are used to process the values and operators
passed by the user. These methods include:
- processNumber(): This method listens for an Action Event on every number and
displays the specific number pressed by the user on the input Label. It also stores
this number in a variable number1.
- processBinaryOperator(): This method also listens for an Action Event on every
operator that takes in two arguments as parameters. It first checks if the operator
is the ‘equal operator’, so that no computation is carried out. Then the operator is
stored into the ‘operator variable’. And the computation of the binary operator is
computed on it’s two operands.
- clearFunction(): This method listens for an Action Event on the ‘Clear’ button,
and it functions in clearing the text on the input and result labels, if another
computation is to be carried out.

The Calculate.java file: This file contains the Calculate class which focuses on how the
operators are used to perform the computations of the expressions. It has 2 methods:
- calculateUnaryNumber(): This method is used on operators which require just one
operand like the square root, sine, cosine, tangent, factorial operators. A Switch
construct is used to check for each operator and the body for each case statement
is used to return the implementation of the operator.
- calculateBinaryNumber(): This method is used on operators which require two
operands like the addition, multiplication, division, subtraction, modulus
operators. A Switch construct is used to check for each operator and the body for
each case statement is used to return the implementation of the operator.

Screenshots of the app are attached.